+++ ftw/addressblock/tests/builders.py
@@ -1,7 +1,5 @@
 from ftw.builder import builder_registry
 from ftw.builder.dexterity import DexterityBuilder
-from ftw.subsite.tests import builders
-from ftw.simplelayout.tests import builders
 
 
 class AddressBlockBuilder(DexterityBuilder):
