+++ ftw/addressblock/tests/test_schema.py
@@ -1,6 +1,6 @@
 from ftw.addressblock.content.addressblock import IAddressBlockSchema
 from ftw.addressblock.interfaces import IAddressBlock
-from ftw.addressblock.tests import FunctionalTestCase
+from ftw.addressblock.tests.base import FunctionalTestCase
 from ftw.builder import Builder
 from ftw.builder import create
 from ftw.testbrowser import browsing
